refactor(theming): Theming stack enhancements
Reduced the number of event listeners attached to the document body.
Theming controllers now register their callbacks into an utiliy object
which handles the theming event dispatch.

Removed the workarounds for `adoptedStyles` since the target browsers
for the library now support the API out of the box.

The theming decorator now exposes (if configured) the theme property from the controller
on the element instance through a private symbol. This reduces the public API
pollution and possible name clashes.