perf(core): Improved LTR-RTL detection
Changed `isLTR` to use the new `:dir(ltr|rtl)` selector,
instead of the `getComputedStyle` call which triggers a style reflow.
After some prelimenary benchmarks it seems to be ~80% faster for repeated calls.

This is already baseline for our target browsers but I had to updated the playwright package
version so it can support the new syntax.
As an added bonus since the new version supports `URL.canParse` I've added
a test for the input as well.