refactor: Expansion panel async hide/show API invokations
In an effort to standardize the public API between every "toggleable"
component, that is any component with an `open` property and
`hide/show/toggle` methods, the "toggleable" methods now
return a boolean promise after any animation(s) have ran to completion.

This PR migrates the expansion panel as well as
the accordion with its `hideAll/showAll` methods.