fix: convertToDate correctly parses time string portions
Added tests for date-time input. The date utilities
class uses the new parse ISO implementation which
returns `null` instead of `Invalid date` for certain
time portions. Adjusted the tests accordingly.
This means that the date-time input additional checks around
the mask value logic could be simplified but that is for
another PR.