fix(forms): Setting validation properties on a pristine non-dirty form (#1638)
* fix(forms): Setting validation properties on a pristine non-dirty form

Setting validation properties on a pristine non-dirty form element
does not apply invalid styles.

Rewritten the synthetic test for invalidation on a single blur
to follow the new behavior of form associated elements.
Added tests for validation slots and removed some which are
not really applicable.
Some code refactoring in the component to follow the coding style
guide.

Closes #1632