feat: Slot controller (#1767)
Added a slot controller implementation that handles `slotchange` events inside the host render root.
It features typed slots, and the ability to query assigned nodes/elements similar to the Lit's `@queryAssigned<foo>` decorators.

Cleaned up a bunch of code around components which use the new controller, as well some dubious ARIA properties in the DOM of some of the components.