refactor: Form associated mixin and validation slots container
Introduced a `_touched` state for form associated components and dropped
the previous `_dirty` flag.

Changed the behavior of how **invalid** visual state is represented for
components. With the current batch of changes, the visual styles are
applied if the internal validity of the component is **false** and one of
the following is applicable:

* the user has interacted with the component through the UI - either
  by focusing, click, selection, typing or other form of keyboard interaction.
* `reportValidity` was invoked on the form associated component.
* The parent form of the form associated component is being submitted.