fix(dialog): `keepOpenOnEscape` should prevent dialog from closing on Escape key press
Chromium based browsers have a default behavior of
firing a `close` event when the Escape key is repeatedly pressed.
Since that event cannot be cancelled, keep the dialog open
when `keepOpenOnEscape` is set to true.

A better fix would be to use the `closedby` property
but unfortunately that is not supported in Safari.

Closes #1983